+++ api/v1alpha1/doc.go
@@ -1,4 +1,4 @@
 // Package v1alpha1 contains API Schema definitions for the chaos v1alpha1 API group.
 // +kubebuilder:object:generate=true
-// +groupName=chaos.chaosplane.io
+// +groupName=chaos.chaosplane.dev
 package v1alpha1