+++ api/v1alpha1/groupversion_info.go
@@ -1,6 +1,6 @@
 // Package v1alpha1 contains API Schema definitions for the chaos v1alpha1 API group.
 // +kubebuilder:object:generate=true
-// +groupName=chaos.chaosplane.io
+// +groupName=chaos.chaosplane.dev
 package v1alpha1
 
 import (
@@ -10,7 +10,7 @@ import (
 
 var (
 	// GroupVersion is group version used to register these objects.
-	GroupVersion = schema.GroupVersion{Group: "chaos.chaosplane.io", Version: "v1alpha1"}
+	GroupVersion = schema.GroupVersion{Group: "chaos.chaosplane.dev", Version: "v1alpha1"}
 
 	// SchemeBuilder is used to add go types to the GroupVersionResource scheme.
 	SchemeBuilder = &scheme.Builder{GroupVersion: GroupVersion}